+++ src/medical/index.js
@@ -52,6 +52,8 @@ const MedicalElement = (props) => {
     backurl,
     footerHidden = false, // 隐藏按钮
     onTabsChange = () => {},
+    submitButtonProps = {}, // 提交按钮的 API 属性
+    draftButtonProps = {}, // 草稿按钮的 API 属性
   } = props;
 
   const [form] = Form.useForm();
@@ -76,7 +78,9 @@ const MedicalElement = (props) => {
     }, []);
 
     form.validateFields().then((values) => {
-      onFinish('submit', outPutFormValues(values, fieldList), activeTabKey);
+      if (isEmptyObject(values)) {
+        onFinish('submit', outPutFormValues(values, fieldList), activeTabKey);
+      }
     }).catch((err) => {
       console.log(err);
     })
@@ -131,23 +135,26 @@ const MedicalElement = (props) => {
               {temp.template.map((item, index) => {
                 const Component = filterEleMapToComponent(ele, item.enName);
 
-                return <Component {...item} index={index} key={item.enName} />;
+                return (
+                  <React.Fragment>
+                    <Component {...item} index={index} key={item.enName} />
+                    <br /><br />
+                  </React.Fragment>
+                );
               })}
             </Tabs.TabPane>
           ))}
         </Tabs>
       </Form>
-      {footerHidden ? null : (
-        <div className="submit_div" hidden={!template[0]}>
-          <Back url={backurl} />
-          <Button type="primary" onClick={onSubmitDraft}>
-            保存草稿
-          </Button>
-          <Button type="primary" onClick={onSubmit}>
-            提交
-          </Button>
-        </div>
-      )}
+      <div className="submit_div" hidden={!template[0] || footerHidden}>
+        <Back url={backurl} />
+        <Button type="primary" {...draftButtonProps} onClick={onSubmitDraft}>
+          保存草稿
+        </Button>
+        <Button type="primary" {...submitButtonProps} onClick={onSubmit}>
+          提交
+        </Button>
+      </div>
     </EleContext.Provider>
   );
 };